chore: moved migrator commands to new folder.
